Correção de float para Bigdecimal na classe ConsumptionService.java

+++ src/main/java/com/nilerbarcelos/ficticiusClean/services/ConsumptionService.java
@@ -1,6 +1,6 @@
 package com.nilerbarcelos.ficticiusClean.services;
 
 public interface ConsumptionService {
-    float computeTotalPrice(float fuelPrice, float consumedFuel);
-    float computeUsage(float distance, float averageConsumption);
+    BigDecimal computeTotalPrice(BigDecimal fuelPrice, BigDecimal consumedFuel);
+    BigDecimal computeUsage(BigDecimal distance, BigDecimal averageConsumption);
 }
